Move the MiniMax-H3 LoRA tests to tests/lora
